+++ CHANGES.md
@@ -33,6 +33,7 @@ Changes in 12.2
 - Send RosarioSIS version in Addon.fnc.php
 - Add Course Period update Teacher action hook in Actions.php & Courses.fnc.php
 - SQL allow Gradebook Grades < 0, not only '-1' (Excused) in EnterEligibility.php, Grades.php, ProgressReports.php, StudentGrades.php & FinalGrades.inc.php
+- Use FinalGradesQtrOrProCalculate() function in GradebookBreakdown.php
 
 Changes in 12.1.3
 -----------------